telegram_bot/requirements.txt
python-telegram-bot==20.7
httpx==0.25.2
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
loguru==0.7.2
aiofiles==23.2.1
# Database (Same as backend)
SQLAlchemy==2.0.23
asyncmy==0.2.8
aiomysql==0.2.0
# Backend dependencies (needed for importing backend modules)
fastapi==0.109.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
redis==5.0.1
phonenumbers==8.13.26
email-validator==2.1.0
